email-service/requirements.txt
sib-api-v3-sdk==7.6.0
Flask==3.0.0
python-dotenv==1.0.0
Flask-CORS==4.0.0
